pcf8523: add bare-bones RTC driver
This is an RTC chip available from e.g. https://www.adafruit.com/product/5189. It is similar to the PCF8563 which already has a driver, but the control registers seem different enough to justify an independent driver.

I have only implemented basic set/get support so far, as that is all I currently need.